# server/docs/swagger.yaml
basePath: /csat/rest
definitions:
  controllers.UpdateAnswerRequest:
    type: object
  models.LoginRequest:
    properties:
      email:
        example: Test1@yahoo.com
        type: string
      password:
        example: Test@123
        type: string
    type: object
  models.Survey:
    type: object
  models.UpdateAnswerRequest:
    properties:
      survey:
        $ref: '#/definitions/schema.Survey'
    type: object
  models.UpdateUserRequest:
    properties:
      negatives:
        type: string
      positives:
        type: string
      rating:
        type: number
      userFeedbackId:
        type: integer
    type: object
  schema.Survey:
    type: object
host: localhost:8000
info:
  contact: {}
  title: CSAT Swagger documentation
  version: "1.0"
paths:
  /api/survey:
    post:
      consumes:
      - application/json
      description: Create Survey
      parameters:
      - description: Create Survey Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.Survey'
      produces:
      - application/json
      responses:
        "200":
          description: 'Survey Createed Successfully" example:"{''message'': ''Survey
            Createed Successfully''}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Invalid request" example:"{''error'': ''Invalid request''}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid" example:"{''error'':
            ''Unauthorized''}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'Internal server error" example:"{''error'': ''Internal server
            error''}'
          schema:
            additionalProperties: true
            type: object
      summary: Create Survey
      tags:
      - Survey
  /api/survey-answers:
    put:
      consumes:
      - application/json
      description: Create Survey
      parameters:
      - description: Create Survey Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.UpdateAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Survey Createed Successfully" example:"{''message'': ''Survey
            Createed Successfully''}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Invalid request" example:"{''error'': ''Invalid request''}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid" example:"{''error'':
            ''Unauthorized''}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'Internal server error" example:"{''error'': ''Internal server
            error''}'
          schema:
            additionalProperties: true
            type: object
      summary: Create Survey
      tags:
      - Survey
  /api/survey-details:
    get:
      consumes:
      - application/json
      description: Retrieve survey details based on Survey ID
      parameters:
      - default: 2
        description: Survey ID (required)
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Survey details retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No user found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get Survey Details
      tags:
      - Survey
  /api/survey-format:
    get:
      consumes:
      - application/json
      description: Retrieve survey format based on ID
      parameters:
      - default: 56
        description: Survey Format ID (optional)
        in: query
        name: surveyFormatID
        type: integer
      - default: 1
        description: Project ID (optional)
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Survey format retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No user found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get Survey format
      tags:
      - Survey
  /api/surveys:
    get:
      consumes:
      - application/json
      description: Retrieve All surveys based on tenant ID
      parameters:
      - default: 101
        description: Tenant ID (required)
        in: query
        name: tenant_id
        required: true
        type: integer
      - description: Survey Format ID (optional)
        in: query
        name: survey_format_id
        type: integer
      - default: 1
        description: Page (required)
        in: query
        name: page
        required: true
        type: integer
      - default: 5
        description: Limit (required)
        in: query
        name: limit
        required: true
        type: integer
      - description: Status (optional)
        in: query
        name: status
        type: string
      - default: CMS
        description: Account Name (optional)
        in: query
        name: accountName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No user found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: All surveys
      tags:
      - Survey
  /api/team-list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of users based on name and project ID
      parameters:
      - description: User name (optional)
        in: query
        name: name
        type: string
      - default: 301
        description: Project ID (required)
        in: query
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User list retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No users found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user list
      tags:
      - Users
  /api/template-details:
    get:
      consumes:
      - application/json
      description: Retrieve template details based on Template ID
      parameters:
      - default: 2
        description: Template ID (required)
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Template details retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No user found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get Template Details
      tags:
      - Templates
  /api/user:
    get:
      consumes:
      - application/json
      description: Retrieve user details based on User ID
      parameters:
      - default: 2
        description: User ID (required)
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No user found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user Details
      tags:
      - Users
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login existing user
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login User
      tags:
      - Users
  /api/userFeedback:
    put:
      consumes:
      - application/json
      description: Update user feedback details
      parameters:
      - description: Update User Feedback Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'User feedback updated successfully" example:"{''message'':
            ''Feedback updated successfully''}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Invalid request" example:"{''error'': ''Invalid request''}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Token is missing or invalid" example:"{''error'':
            ''Unauthorized''}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'Internal server error" example:"{''error'': ''Internal server
            error''}'
          schema:
            additionalProperties: true
            type: object
      summary: Update user feedback
      tags:
      - Users
swagger: "2.0"